backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
aiohttp
certifi
pydub
slack_bolt
supabase
httpx
asyncio
datetime
openai
google-cloud-pubsub
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
numpy<2
